Move get_concrete_libfunc_id to utils.rs.
commit-id:3e06363d
